feat: move out from ui to it's own component
