+++ src/component/mod.rs
@@ -1,0 +1,5 @@
+pub mod requestbar;
+pub mod responsebar;
+pub mod sidebar;
+pub mod tabbar;
+pub mod urlbar;
