+++ src/component/responsebar.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Default)]
+pub struct ResponseBar {
+    pub body: String,
+}
