+++ src/component/tabbar.rs
@@ -1,0 +1,7 @@
+use crate::items::Item;
+
+#[derive(Debug, Default)]
+pub struct TabBar {
+    pub selected: usize,
+    pub tabs: Vec<Item>,
+}
